requirements.txt
aiofiles==24.1.0
aiogram==3.16.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
arrow==1.3.0
asgiref==3.8.1
attrs==24.3.0
blessed==1.20.0
certifi==2024.12.14
charset-normalizer==3.4.1
Django==3.2.21
django-picklefield==3.2
django-q==1.3.9
frozenlist==1.5.0
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
propcache==0.2.1
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==3.5.3
requests==2.32.3
setuptools==75.6.0
six==1.17.0
sqlparse==0.5.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
yarl==1.18.3
